requirements.txt
mcp[cli]
parallel-web
pydantic
pydantic-settings
pytest
pytest-asyncio
pytest-cov
python-dotenv
requests
ruff
openai